Add the problem description.
For reasons unknown, Infi decided to hide exercise once the solution
was given. Hence, we went to the infi site again, but this time using
incognito mode. That gave us the problem description of the first
part. We downloaded the next problem set, solve it, and got access
to the problem description of the second part.